Style : read와 command 테스트 분리
